Add expiry to all data added to redis
To reduce the amount of data stored in redis and for how long it is
stored this adds an expiry to all data entered into redis by this gem.

By default a candidate results will be stored for 60 seconds and the
results are stored for 24 hours.

This can be customised with the hash argument to `experiment_control`

eg `experiment_control(..., results_expiry: 60 * 5, candidate_expiry: 60)`