Hash results so they are not stored in redis twice
Before data of a mismatch is entered into redis a SHA2 hash is
generated for the data and redis is checked to see if that data already
exists. If it exists the new data is not added and the expiry time is
extended. If it does not exist it is added to redis.